Fix a bug spotted by spoon-b (error when renaming a profile with no log file). Thanks!
git-svn-id: https://synchronicity.svn.sourceforge.net/svnroot/synchronicity/trunk@247 6fc22aba-598d-424d-8c45-4c47160ecdfc

+++ Create Synchronicity/SettingsHandling.vb
@@ -286,7 +286,7 @@ Class ProfileHandler
         If IO.File.Exists(ProgramConfig.GetLogPath(NewName)) Or IO.File.Exists(ProgramConfig.GetConfigPath(NewName)) Then Return False
 
         Try
-            IO.File.Move(ProgramConfig.GetLogPath(ProfileName), ProgramConfig.GetLogPath(NewName))
+            If IO.File.Exists(ProgramConfig.GetLogPath(ProfileName)) Then IO.File.Move(ProgramConfig.GetLogPath(ProfileName), ProgramConfig.GetLogPath(NewName))
             IO.File.Move(ProgramConfig.GetConfigPath(ProfileName), ProgramConfig.GetConfigPath(NewName))
         Catch
             Return False
